[COMMON] fimc-is2: change vendor_luxStandard to float
CRs-fixed: (CR)

Reviewed-on: https://gerrit.mot.com/1324233
SLTApproved: Slta Waiver
SME-Granted: SME Approvals Granted
Tested-by: Jira Key
Submit-Approved: Jira Key

+++ drivers/media/platform/exynos/fimc-is2/include/fimc-is-metadata.h
@@ -1127,7 +1127,7 @@ struct camera2_aa_dm {
 	uint32_t			vendor_fusionCaptureAfInfo;
 	uint32_t			vendor_colorTempIndex;
 	uint32_t			vendor_luxIndex;
-	uint32_t			vendor_luxStandard;
+	float				vendor_luxStandard;
 	uint32_t			vendor_reserved[6];
 };
